Feature/dry run

+++ main.py
@@ -18,8 +18,7 @@ def calculate_checksum(file_path):
             hasher.update(chunk)
     return hasher.hexdigest()
 
-
-def upload_to_s3(local_dir, s3_bucket, s3_prefix, exclude_list):
+def upload_to_s3(local_dir, s3_bucket, s3_prefix, exclude_list, dry_run=False):
     s3 = boto3.client('s3')
 
     for root, dirs, files in os.walk(local_dir):
@@ -29,10 +28,14 @@ def upload_to_s3(local_dir, s3_bucket, s3_prefix, exclude_list):
                 local_path = os.path.join(root, file)
                 relative_path = os.path.relpath(local_path, local_dir)
                 s3_key = os.path.join(s3_prefix, relative_path)
-                print(f"Uploading {file} to S3 bucket {s3_bucket}")
-                s3.upload_file(local_path, s3_bucket, s3_key)
+                
+                if dry_run:
+                    print(f"Simulating: Would upload {file} to S3 bucket {s3_bucket} as {s3_key}")
+                else:
+                    print(f"Uploading {file} to S3 bucket {s3_bucket}")
+                    s3.upload_file(local_path, s3_bucket, s3_key)
 
-def download_from_s3(s3_bucket, s3_prefix, local_dir, exclude_list):
+def download_from_s3(s3_bucket, s3_prefix, local_dir, exclude_list, dry_run=False):
     s3 = boto3.client('s3')
 
     response = s3.list_objects_v2(Bucket=s3_bucket, Prefix=s3_prefix)
@@ -40,15 +43,19 @@ def download_from_s3(s3_bucket, s3_prefix, local_dir, exclude_list):
         s3_key = obj['Key']
         if not any(item in s3_key for item in exclude_list):
             local_path = os.path.join(local_dir, os.path.relpath(s3_key, s3_prefix))
-            print(f"Downloading {s3_key} from S3 bucket {s3_bucket}")
-            os.makedirs(os.path.dirname(local_path), exist_ok=True)
-            s3.download_file(s3_bucket, s3_key, local_path)
-
+            
+            if dry_run:
+                print(f"Simulating: Would download {s3_key} from S3 bucket {s3_bucket} to {local_path}")
+            else:
+                print(f"Downloading {s3_key} from S3 bucket {s3_bucket}")
+                os.makedirs(os.path.dirname(local_path), exist_ok=True)
+                s3.download_file(s3_bucket, s3_key, local_path)
 
 def main():
     parser = argparse.ArgumentParser(description="Upload and download directories/files from Amazon S3")
     parser.add_argument("--upload", help="Upload directories/files to S3", action="store_true")
     parser.add_argument("--download", help="Download directories/files from S3", action="store_true")
+    parser.add_argument("--dry-run", help="Simulate the operation without actual upload/download", action="store_true")
     parser.add_argument("--exclude", nargs='*', help="Exclude files or directories from upload/download", default=[])
     args = parser.parse_args()
 
@@ -72,12 +79,14 @@ def main():
     exclude_list = args.exclude + ignored_items + ['main.py', '.config.ini', '.git']
 
     if args.upload:
-        upload_to_s3(local_dir, s3_bucket, s3_prefix, exclude_list)
-        print("Upload process completed.")
+        upload_to_s3(local_dir, s3_bucket, s3_prefix, exclude_list, args.dry_run)
+        if not args.dry_run:
+            print("Upload process completed.")
 
     elif args.download:
-        download_from_s3(s3_bucket, s3_prefix, local_dir,  exclude_list)
-        print("Download process completed.")
+        download_from_s3(s3_bucket, s3_prefix, local_dir, exclude_list, args.dry_run)
+        if not args.dry_run:
+            print("Download process completed.")
 
 if __name__ == "__main__":
     main()
